refactor into bench_bandwidth.rs
